break FlashBanner into multiple files
